fix: move diagnostics finding interface to dedicated file
Agent-Logs-Url: https://github.com/hiero-ledger/solo/sessions/1d401565-0f0e-49ac-95b7-46e513b81a94
